# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UserController_getAllUsers","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UserController_getUserById","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_doRegister","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_doLogin","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/categories":{"get":{"operationId":"CategoryController_getAll","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Categories"],"security":[{"bearer":[]}]},"post":{"operationId":"CategoryController_create","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["Categories"],"security":[{"bearer":[]}]}},"/categories/{id}":{"get":{"operationId":"CategoryController_findOneById","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Categories"],"security":[{"bearer":[]}]},"patch":{"operationId":"CategoryController_update","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["Categories"],"security":[{"bearer":[]}]},"delete":{"operationId":"CategoryController_destroy","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Categories"],"security":[{"bearer":[]}]}},"/todos":{"get":{"operationId":"TodoController_findAll","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Todo"],"security":[{"bearer":[]}]},"post":{"operationId":"TodoController_create","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDTO"}}}},"responses":{"201":{"description":""}},"tags":["Todo"],"security":[{"bearer":[]}]}},"/todos/{id}":{"get":{"operationId":"TodoController_findOneById","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Todo"],"security":[{"bearer":[]}]},"patch":{"operationId":"TodoController_update","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDTO"}}}},"responses":{"200":{"description":""}},"tags":["Todo"],"security":[{"bearer":[]}]},"delete":{"operationId":"TodoController_destroy","parameters":[{"name":"x-api-version","in":"header","description":"API version. ieg: 1","schema":{"type":"string"}},{"name":"x-lang","in":"header","description":"User Language. ieg: id","schema":{"type":"string"}},{"name":"x-timezone","in":"header","description":"User timezone. ieg: Asia/Jakarta","schema":{"type":"string"}},{"name":"x-version","in":"header","description":"Client app version ieg: 1.0.0","schema":{"type":"string"}},{"name":"x-device","in":"header","description":"Client app platform ios or android","schema":{"type":"string"}},{"name":"x-user-id","in":"header","description":"Logged in user id","schema":{"type":"string"}},{"name":"x-user-uuid","in":"header","description":"Logged in user uuid","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Todo"],"security":[{"bearer":[]}]}}},"info":{"title":"Talentclass API ","description":"Talentclass API Documentation","version":"1.0","contact":{"name":"Sociolite","url":"https://sociolite.id","email":"hello@sociolite.id"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"}},"tags":[],"servers":[{"url":"/api"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthRegisterDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"john.doe@example.com"},"password":{"type":"string","example":"password123"}},"required":["name","email","password"]},"AuthLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"CreateTodoDTO":{"type":"object","properties":{"category_id":{"type":"integer"},"title":{"type":"string"},"description":{"type":"string"},"completed":{"type":"boolean"}},"required":["category_id","title","description","completed"]}}}}